Implementação do repositório de MemberRepository: funcionalidades de criar, buscar, atualizar e deletar membros adicionadas
